+++ packages/@runlightyear/airtable/src/index.ts
@@ -1,0 +1,4 @@
+import { Airtable } from "./Airtable";
+import { AirtableOauth } from "./AirtableOauth";
+
+export { Airtable, AirtableOauth };