+++ packages/@runlightyear/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/cli
 
+## 0.16.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.16.0
+
 ## 0.16.0
 
 ### Minor Changes