+++ packages/@runlightyear/openai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/openai
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.16.0
+
 ## 0.1.1
 
 ### Patch Changes